Dataview

+++ src/ts/component/page/index.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { RouteComponentProps } from 'react-router';
 import { Util, Storage, analytics, keyboard } from 'ts/lib';
+import { ListPopup } from 'ts/component';
 import { authStore, commonStore } from 'ts/store';
-import { observer } from 'mobx-react';
 
 import PageAuthInvite from './auth/invite';
 import PageAuthNotice from './auth/notice';
@@ -50,6 +50,8 @@ const Components: any = {
 
 interface Props extends RouteComponentProps<any> {
 	isPopup?: boolean;
+	matchPopup?: any;
+	rootId?: string;
 };
 
 class Page extends React.Component<Props, {}> {
@@ -58,7 +60,8 @@ class Page extends React.Component<Props, {}> {
 	childRef: any;
 
 	render () {
-		const { match } = this.props;
+		const { isPopup } = this.props;
+		const match = this.getMatch();
 		const path = [ match.params.page, match.params.action ].join('/');
 		const showNotice = !Boolean(Storage.get('firstRun'));
 		
@@ -73,9 +76,12 @@ class Page extends React.Component<Props, {}> {
 		};
 		
 		return (
-			<div className={'page ' + this.getClass('page')}>
-				<Component ref={(ref: any) => this.childRef = ref} {...this.props} />
-			</div>
+			<React.Fragment>
+				{!isPopup ? <ListPopup {...this.props} /> : ''}
+				<div className={'page ' + this.getClass('page')}>
+					<Component ref={(ref: any) => this.childRef = ref} {...this.props} />
+				</div>
+			</React.Fragment>
 		);
 	};
 	
@@ -92,10 +98,16 @@ class Page extends React.Component<Props, {}> {
 		this._isMounted = false;
 		this.unbind();
 	};
-	
+
+	getMatch () {
+		const { match, matchPopup, isPopup } = this.props;
+		return isPopup ? matchPopup : match;
+	};
+
 	init () {
 		const { account } = authStore;
-		const { isPopup, match, history } = this.props;
+		const { isPopup, history } = this.props;
+		const match = this.getMatch();
 		const popupNewBlock = Storage.get('popupNewBlock');
 		const isIndex = !match.params.page;
 		const isAuth = match.params.page == 'auth';
@@ -118,14 +130,8 @@ class Page extends React.Component<Props, {}> {
 		this.event();
 		this.unbind();
 
-		if (!isPopup) {
-			commonStore.popupCloseAll();
-			commonStore.menuCloseAll();
-		};
-
 		Util.linkPreviewHide(true);
 		keyboard.setMatch(match);
-
 		
 		if (isMain && !popupNewBlock) {
 			commonStore.popupOpen('help', { 
@@ -133,31 +139,37 @@ class Page extends React.Component<Props, {}> {
 			});
 		};
 
-		if (isMainIndex) {
-			if (account && askSurvey && !commonStore.popupIsOpen() && !lastSurveyCanceled && (lastSurveyTime <= Util.time() - 86400 * days)) {
-				Storage.delete('askSurvey');
-				commonStore.popupOpen('confirm', {
-					data: {
-						title: 'We need your opinion',
-						text: 'Please, tell us what you think about Anytype. Participate in 1 min survey',
-						textConfirm: 'Let\'s go!',
-						textCancel: 'Skip',
-						canCancel: true,
-						onConfirm: () => {
-							ipcRenderer.send('urlOpen', Util.sprintf(Constant.survey, account.id));
-							Storage.set('lastSurveyTime', Util.time());
-						},
-						onCancel: () => {
-							Storage.set('lastSurveyCanceled', 1);
-							Storage.set('lastSurveyTime', Util.time());
-						},
-					},
-				});
-			};
+		if (!isPopup) {
+			commonStore.popupCloseAll();
+			commonStore.menuCloseAll();
 
-			Storage.delete('redirect');
-		} else {
-			Storage.set('redirect', history.location.pathname);
+			if (isMainIndex) {
+				if (account && askSurvey && !commonStore.popupIsOpen() && !lastSurveyCanceled && (lastSurveyTime <= Util.time() - 86400 * days)) {
+					Storage.delete('askSurvey');
+
+					commonStore.popupOpen('confirm', {
+						data: {
+							title: 'We need your opinion',
+							text: 'Please, tell us what you think about Anytype. Participate in 1 min survey',
+							textConfirm: 'Let\'s go!',
+							textCancel: 'Skip',
+							canCancel: true,
+							onConfirm: () => {
+								ipcRenderer.send('urlOpen', Util.sprintf(Constant.survey, account.id));
+								Storage.set('lastSurveyTime', Util.time());
+							},
+							onCancel: () => {
+								Storage.set('lastSurveyCanceled', 1);
+								Storage.set('lastSurveyTime', Util.time());
+							},
+						},
+					});
+				};
+
+				Storage.delete('redirect');
+			} else {
+				Storage.set('redirect', history.location.pathname);
+			};
 		};
 
 		$(window).on('resize.page', () => { this.resize(); });
@@ -168,7 +180,7 @@ class Page extends React.Component<Props, {}> {
 	};
 	
 	event () {
-		const { match } = this.props;
+		const match = this.getMatch();
 		const page = String(match.params.page || 'index');
 		const action = String(match.params.action || 'index');
 		const path = [ 'page', page, action ].join('-');
@@ -177,7 +189,7 @@ class Page extends React.Component<Props, {}> {
 	};
 	
 	getClass (prefix: string) {
-		const { match } = this.props;
+		const match = this.getMatch();
 		const page = match.params.page || 'index';
 		const action = match.params.action || 'index';
 		const platform = Util.getPlatform();
@@ -197,6 +209,9 @@ class Page extends React.Component<Props, {}> {
 		
 		if (config.debug.ui) {
 			cn.push('debug');
+		};
+		if (config.debug.dm) {
+			cn.push('dark');
 		};
 
 		obj.attr({ class: cn.join(' ') });
